requirements-dev.txt
pytest>=7.0
pytest-cov>=4.1
pyyaml>=6.0
matplotlib>=3.7
jsonschema>=4.20
scikit-learn>=1.3
fastapi>=0.105
uvicorn>=0.23
mediapipe>=0.10
ultralytics>=8.0
gaze-lle @ git+https://github.com/lllyasviel/GazeLL-E.git

